Fixed History Instantiation (WIP) #12 #11 #9
Some variadic functions pass the variadic filter and create a mismatch in number of args and params,
resulting in illegal vector/map references in historyContext/params_
vector.

History Represenstation needs to be integrated with the cast and var matchers to build complete
history for an operand. Some fixes may be needed for LinkedParmQn or how the arg key is
calculated to better link histories.