Local Context Separated from History (WIP) #13
Local/global history differentiation is clearer with local context, i.e. context specific to a call/line
is stored separately. This allows history references to be kept in branches which then allows to keep track of
updates in global history.

History key resolution using context now tries to match key in TypeTransforms if resolved key is not found in
census. This allows better fptr resolution.