Fixed single use void + Bitcast check (#40)
Instead of checking just in/out score, in/out types are also checked to
ensure that void * is only used with a single type.

Some bitcasts were missed because the matcher and subsequent code only
check for the top-level casts. If a bitcast is nested:
    void *pv;
    int i = *(int*)pv;
then it is missed and only LtoR cast is recorded. This not only fixes
some stats it also improves detection because pattern detection often
checks if a summary link is from a bitcast.