Add count for pointers that are never used in cast (#22)
This count includes all pointers, regardless of whether they are void or not, fptr param
or not. Function pointers are not included.

Also improved how stats are displayed. Now summarized data from both type summary and Stat match
are displayed at the end in a single block with necessary tables.